Add more custom spacy pipes

+++ dev_env.bash
@@ -1,11 +1,14 @@
 #!/usr/bin/env bash
 
-deactivate
+if [[ ! -z "$VIRTUAL_ENV" ]]; then
+  echo "'deactivate' before running this script."
+  exit 1
+fi
 
 rm -rf venv
 virtualenv -p python3.8 venv
 
-./venv/bin/activate
+source ./venv/bin/activate
 
 python -m pip install --upgrade pip setuptools wheel
 if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
